Create CollapsedStory component
The StoryItem was divided in a different file named CollapsedStory
to be the abse of the future features.
